move utils.util into init for now, it's a good place for it
